+++ scanpipe/tests/pipes/test_output.py
@@ -267,6 +267,31 @@ class ScanPipeOutputPipesTest(TestCase):
         output_file = output.to_spdx(project=project)
         self.assertIn(output_file.name, project.output_root)
 
+    def test_scanpipe_pipes_outputs_to_spdx_extracted_licenses(self):
+        project = Project.objects.create(name="Analysis")
+        package_data = dict(package_data1)
+        # ac3filter resolves as LicenseRef-scancode-ac3filter
+        expression = "mit AND ac3filter"
+        package_data["declared_license_expression"] = expression
+        pipes.update_or_create_package(project, package_data)
+
+        output_file = output.to_spdx(project=project, include_files=True)
+        self.assertIn(output_file.name, project.output_root)
+
+        results_json = json.loads(output_file.read_text())
+        # mit is part of the SPDX license list, thus not in hasExtractedLicensingInfos
+        self.assertEqual(1, len(results_json["hasExtractedLicensingInfos"]))
+        license_infos = results_json["hasExtractedLicensingInfos"][0]
+        self.assertEqual("LicenseRef-scancode-ac3filter", license_infos["licenseId"])
+        self.assertEqual("AC3Filter License", license_infos["name"])
+        expected = [
+            "http://www.ac3filter.net/wiki/Download_AC3Filter",
+            "http://ac3filter.net",
+            "http://ac3filter.net/forum",
+        ]
+        self.assertEqual(expected, license_infos["seeAlsos"])
+        self.assertTrue(license_infos["extractedText"].startswith("License:"))
+
     def test_scanpipe_pipes_outputs_make_unknown_license_object(self):
         licensing = get_licensing()
         parsed_expression = licensing.parse("some-unknown-license")